// src/app/helpers/RoomCell.enums.ts
export enum RoomCellPublicEvents {
  userJoins = 'userJoins',
  userLeaves = 'userLeaves',
  userUpdates = 'userUpdates',

  userJoinsCall = 'userJoinsCall',
  userLeavesCall = 'userLeavesCall',

  newMessage = 'newMessage',

  userAddsVideoStream = 'userAddsVideoStream',
  userRemovesVideoStream = 'userRemovesVideoStream',

  recvVideoStream = 'recvVideoStream',
  remvVideoStream = 'remvVideoStream',

  signal = 'signal',
}